+++ test/module/domain/product/product.controller.spec.ts
@@ -2,27 +2,27 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { MongooseModule } from '@nestjs/mongoose';
 import { NestFastifyApplication } from '@nestjs/platform-fastify';
 import { ConfigModule } from '@nestjs/config';
-import { MarketRepository } from '../../../../domain/market/market.repository';
-import { MarketModule } from '../../../../domain/market/market.module';
-import { UserRepository } from '../../../../domain/user/user.repository';
-import { AuthModule } from '../../../../module/auth/auth.module';
-import { validationPipe } from '../../../../common/pipe/validation.pipe';
-import { sessionConfig } from '../../../../config/session.config';
-import { ProductModule } from '../../../../domain/product/product.module';
-import { ProductRepository } from '../../../../domain/product/product.repository';
 import * as request from 'supertest';
 import { Types } from 'mongoose';
 import * as bcrypt from 'bcrypt';
-import { User } from '../../../../domain/user/user.schema';
-import { Market } from '../../../../domain/market/market.schema';
-import { MarketErrorMessage } from '../../../../domain/market/market.message';
-import { ProductErrorMessage } from '../../../../domain/product/product.message';
-import { Country } from '../../../../domain/common/enums/Country';
-import { Category } from '../../../../domain/common/enums/Category';
-import { CommonErrorMessage } from '../../../../common/error/common.message';
-import { Product } from '../../../../domain/product/product.schema';
-import { localDateTimeToString } from '../../../../common/util/dateUtil';
 import { HttpStatus } from '@nestjs/common';
+import { UserRepository } from '../../../../src/domain/user/user.repository';
+import { MarketRepository } from '../../../../src/domain/market/market.repository';
+import { ProductRepository } from '../../../../src/domain/product/product.repository';
+import { User } from '../../../../src/domain/user/user.schema';
+import { Market } from '../../../../src/domain/market/market.schema';
+import { AuthModule } from '../../../../src/module/auth/auth.module';
+import { MarketModule } from '../../../../src/domain/market/market.module';
+import { ProductModule } from '../../../../src/domain/product/product.module';
+import { validationPipe } from '../../../../src/common/pipe/validation.pipe';
+import { sessionConfig } from '../../../../src/config/session.config';
+import { Country } from '../../../../src/domain/common/enums/Country';
+import { Category } from '../../../../src/domain/common/enums/Category';
+import { MarketErrorMessage } from '../../../../src/domain/market/market.message';
+import { ProductErrorMessage } from '../../../../src/domain/product/product.message';
+import { CommonErrorMessage } from '../../../../src/common/error/common.message';
+import { Product } from '../../../../src/domain/product/product.schema';
+import { localDateTimeToString } from '../../../../src/common/util/dateUtil';
 
 describe('ProductController', () => {
   let app: NestFastifyApplication;
