+++ test/module/domain/product/product.repository.spec.ts
@@ -2,20 +2,20 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { NestFastifyApplication } from '@nestjs/platform-fastify';
 import { ConfigModule } from '@nestjs/config';
 import { MongooseModule } from '@nestjs/mongoose';
-import { ProductRepository } from '../../../../domain/product/product.repository';
-import { SearchProductsDto } from '../../../../domain/product/product.request.dto';
-import { ProductModule } from '../../../../domain/product/product.module';
-import { UserModule } from '../../../../domain/user/user.module';
-import { MarketModule } from '../../../../domain/market/market.module';
-import { UserRepository } from '../../../../domain/user/user.repository';
-import { MarketRepository } from '../../../../domain/market/market.repository';
 import * as bcrypt from 'bcrypt';
-import { User } from '../../../../domain/user/user.schema';
-import { Country } from '../../../../domain/common/enums/Country';
-import { Market } from '../../../../domain/market/market.schema';
-import { Category } from '../../../../domain/common/enums/Category';
-import { Product } from '../../../../domain/product/product.schema';
 import { Types } from 'mongoose';
+import { UserRepository } from '../../../../src/domain/user/user.repository';
+import { MarketRepository } from '../../../../src/domain/market/market.repository';
+import { ProductRepository } from '../../../../src/domain/product/product.repository';
+import { Market } from '../../../../src/domain/market/market.schema';
+import { User } from '../../../../src/domain/user/user.schema';
+import { UserModule } from '../../../../src/domain/user/user.module';
+import { ProductModule } from '../../../../src/domain/product/product.module';
+import { MarketModule } from '../../../../src/domain/market/market.module';
+import { Country } from '../../../../src/domain/common/enums/Country';
+import { Category } from '../../../../src/domain/common/enums/Category';
+import { Product } from '../../../../src/domain/product/product.schema';
+import { SearchProductsDto } from '../../../../src/domain/product/product.request.dto';
 
 describe('ProductRepository', () => {
   let app: NestFastifyApplication;
